Add support for localnet

+++ init.sh
@@ -11,6 +11,10 @@ data="$(realpath ./data)/node0"
 #accountid="zentria.poolv1.near"
 #data="$(realpath ./mainnet_data)/node0"
 
+#network="localnet"
+#accountid=""
+#data="$(realpath ./localnet_data)/node0"
+
 neard="$(realpath -- ./result/bin/neard)"
 config_url="https://s3-us-west-1.amazonaws.com/build.nearprotocol.com/nearcore-deploy/${network}/config.json"
 genesis_url="https://s3-us-west-1.amazonaws.com/build.nearprotocol.com/nearcore-deploy/${network}/genesis.json"
@@ -23,17 +27,21 @@ if ! [ -d "${data}" ]; then
 	fi
 
 	# Set up genesis and config
-	mkdir -p "${data}" "${data}/data"
-	curl -o "${data}/genesis.json" "${genesis_url}"
-	curl -o "${data}/config.json" "${config_url}"
+	if ! [ "${network}" = "localnet" ]; then
+		mkdir -p "${data}" "${data}/data"
+		curl -o "${data}/genesis.json" "${genesis_url}"
+		curl -o "${data}/config.json" "${config_url}"
+	fi
 
 	# Initialize data dir
 	"${neard}" --home "${data}" init "${init_args[@]}"
 
 	# Download database dump to speed up syncing
-	aria2c --dir="${data}" --continue=true --max-connection-per-server=16 --lowest-speed-limit=10M --max-tries=2147483647 "${dump_url}"
-	pv "${data}/data.tar" | tar -C "${data}/data" -xf -
-	rm "${data}/data.tar"
+	if ! [ "${network}" = "localnet" ]; then
+		aria2c --dir="${data}" --continue=true --max-connection-per-server=16 --lowest-speed-limit=10M --max-tries=2147483647 "${dump_url}"
+		pv "${data}/data.tar" | tar -C "${data}/data" -xf -
+		rm "${data}/data.tar"
+	fi
 fi
 
 systemd-run --user --pty --unit="neard-${network}" "${neard}" --home "${data}" run
